+++ src/main/java/com/app/concessionario/services/AutoService.java
@@ -50,18 +50,46 @@ public class AutoService {
 
 
 //    chiamata post per creare una nuova auto
-//    ###### TODO gestire creazione accessorio con stesso id
-    public void addAutoDTO(AutoDTO autoDTO) {
-        List<Motore> motori = motoreRepository.findAll();
-        List<Accessorio> accessori = accessorioRepository.findAll();
-        List<Cliente> clienti = clienteRepository.findAll();
-        List<Concessionario> concessionari = concessionarioRepository.findAll();
-        Auto newAuto = AutoMapper.toEntity(autoDTO, motori, accessori, clienti, concessionari);
-        autoRepository.save(newAuto);
+    public void addAutoDTO(AutoDTO autoDTO) throws Exception {
+        List<Integer> autoIds = autoRepository.findAll().stream().map(Auto::getId).toList();
+        if (autoIds.contains(autoDTO.getId())) {
+            throw new Exception("L'id inserito non é valido");
+        } else if (autoDTO.getBrand() == null) {
+            throw new Exception("Il campo brand non può essere vuoto");
+        } else if (autoDTO.getModello() == null) {
+            throw new Exception("Il campo modello non può essere vuoto");
+        } else if (autoDTO.getBrand().length() > 50) {
+            throw new Exception("La lunghezza del campo brand non può superare i 50 caratteri");
+        } else if (autoDTO.getModello().length() > 50) {
+            throw new Exception("La lunghezza del campo modello non può superare i 50 caratteri");
+        } else if (autoDTO.getColore().length() > 20) {
+            throw new Exception("La lunghezza del campo colore non può superare i 20 caratteri");
+        } else if (autoDTO.getStato() != null && autoDTO.getStato().length() > 50) {
+            throw new Exception("La lunghezza del campo stato non può superare i 20 caratteri");
+        } else if (autoDTO.getCarrozzeria() == null) {
+            throw new Exception("Il campo carrozzeria non può essere vuoto");
+        } else if (autoDTO.getCarrozzeria() != null) {
+            if (autoDTO.getCarrozzeria().equalsIgnoreCase("suv") ||
+                    autoDTO.getCarrozzeria().equalsIgnoreCase("berlina") ||
+                    autoDTO.getCarrozzeria().equalsIgnoreCase("crossover") ||
+                    autoDTO.getCarrozzeria().equalsIgnoreCase("stationwagen") ||
+                    autoDTO.getCarrozzeria().equalsIgnoreCase("utilitaria")) {
+
+                List<Motore> motori = motoreRepository.findAll();
+                List<Accessorio> accessori = accessorioRepository.findAll();
+                List<Cliente> clienti = clienteRepository.findAll();
+                List<Concessionario> concessionari = concessionarioRepository.findAll();
+                Auto newAuto = AutoMapper.toEntity(autoDTO, motori, accessori, clienti, concessionari);
+                autoRepository.save(newAuto);
+            } else {
+                throw new Exception("I campi carrozzeria disponibili sono: 'suv', 'berlina', 'stationwagon', 'crossover', 'utilitaria'");
+            }
+
+        }
     }
 
 //    chiamata put per update auto
-    public void updateAutoDTO(Integer id, AutoDTO autoDTO) {
+    public void updateAutoDTO(Integer id, AutoDTO autoDTO) throws Exception {
         List<Motore> motori = motoreRepository.findAll();
         List<Accessorio> accessori = accessorioRepository.findAll();
         List<Cliente> clienti = clienteRepository.findAll();
@@ -72,23 +100,35 @@ public class AutoService {
     }
 
 //    chiamata per eliminare un auto
-//    ##### TODO GESTIRE ID NON ESISTENTE
-    public void deleteAuto(Integer id) {
-    autoRepository.deleteById(id);
+    public void deleteAuto(Integer id) throws Exception {
+        List<Integer> autoIds = autoRepository.findAll().stream().map(Auto::getId).toList();
+        if (autoIds.contains(id)) {
+            autoRepository.deleteById(id);
+        }
+        throw new Exception("L'id inserito non é valido");
 }
 
 
 
 //    METODO PER AGGIORNARE IL CLIENTE ID DI UN AUTO QUANDO VIENE VENDUTA
-    public void sellAuto(Integer autoId, Integer clienteId) {
-        Optional<Auto> a = autoRepository.findById(autoId);
-        Optional<Cliente> c = clienteRepository.findById(clienteId);
-                Auto auto = a.get();
-                Cliente cliente = c.get();
-                auto.setCliente(cliente);
-                auto.setIsVentuta(true);
-                autoRepository.save(auto);
-
+    public void sellAuto(Integer autoId, Integer clienteId) throws Exception {
+        List<Integer> autoIds = autoRepository.findAll().stream().map(Auto::getId).toList();
+        List<Integer> clientiIds = clienteRepository.findAll().stream().map(Cliente::getId).toList();
+            Optional<Auto> a = autoRepository.findById(autoId);
+            Optional<Cliente> c = clienteRepository.findById(clienteId);
+            Auto auto = a.get();
+            Cliente cliente = c.get();
+        if (!autoIds.contains(autoId)) {
+            throw new Exception("L'id dell'auto inserita non è presente");
+        } else if (auto.getIsVentuta()) {
+            throw new Exception("L'auto selezionata ha giá un cliente");
+        } else if (!clientiIds.contains(clienteId)) {
+            throw new Exception("L'id del cliente inserito non è presente");
+        } else {
+            auto.setCliente(cliente);
+            auto.setIsVentuta(true);
+            autoRepository.save(auto);
+        }
 
 
     }
